Update ruby dependencies for Ruby >= 2.2.
Previous version did not specify "rspec" in the Gemfile, which caused `rake`
(as specified in TESTING.md) to not work after a `bundle install`.

Previous version pinned to json 1.7.6, which is incompatible with Ruby >= 2.2.
New version is autogenerated as of 2015-04-09 and works with the supplied rake
tasks.

Previous version used `:gemcutter` for the gems source which uses http. This
was switched to 'https://rubygems.org' as recommended by bundler docs.