Add Readonly types
This commit adds Readonly types and changes many of the function
signatures to use them in appropriate places.

Docs and dist has not been updated, because I am not 100% confident
that these are the final changes and I don't want to bloat history with
tons of unnecessary file changes until the work is complete.